rewrite differential test
